Work on Motors and Muscles, and a few figure captions here and there.
I made a few figure captions toward the end of the Optimization chapter, and I did some major revisions on the Motors and Muscles chapter. I also added one reference (Ogata) to the bibliography. I use bibdesk, which changed the formatting of the .bib file a little bit, but not in an incomprehensible way. The reference (Ogata) was added to Passive Dynamics chapter.